ingestion_service/requirements.txt
psycopg2-binary==2.9.9
confluent-kafka==2.3.0
minio==7.2.16
PyYAML==6.0.1
pydantic==2.8.2
pytest==7.4.0
pytest-mock==3.12.0